@@ -3,19 +3,28 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
+    "start": "react-native start",
+    "test": "jest",
+    "lint": "eslint .",
+    "pod-install": "cd ios && pod install"
   },
   "dependencies": {
-    "react": "16.4.1",
-    "react-native": "0.56.0",
-    "react-native-push-notification": "git+https://git@github.com/zo0r/react-native-push-notification.git"
+    "@react-native-community/push-notification-ios": "^1.10.1",
+    "react": "16.13.1",
+    "react-native": "0.63.3",
+    "react-native-push-notification": "zo0r/react-native-push-notification#dev"
   },
   "devDependencies": {
-    "babel-jest": "23.4.0",
-    "babel-preset-react-native": "^5",
-    "jest": "23.4.1",
-    "react-test-renderer": "16.4.1"
+    "@babel/core": "^7.9.0",
+    "@babel/runtime": "^7.9.2",
+    "@react-native-community/eslint-config": "^1.0.0",
+    "babel-jest": "^25.3.0",
+    "eslint": "^6.8.0",
+    "jest": "^25.3.0",
+    "metro-react-native-babel-preset": "^0.59.0",
+    "react-test-renderer": "16.11.0"
   },
   "jest": {
     "preset": "react-native"
