@@ -8,9 +8,9 @@ import android.os.Bundle;
 import android.util.Log;
 
 class RNPushNotificationConfig {
-    private static final String KEY_CHANNEL_PREFIX = "com.dieam.reactnativepushnotification.channel_prefix";
-    private static final String KEY_CHANNEL_NAME = "com.dieam.reactnativepushnotification.notification_channel_name";
-    private static final String KEY_CHANNEL_DESCRIPTION = "com.dieam.reactnativepushnotification.notification_channel_description";
+    private static final String KEY_NOTIFICATION_FIREBASE_DEFAULT_CHANNEL_ID = "com.google.firebase.messaging.default_notification_channel_id";
+    private static final String KEY_NOTIFICATION_DEFAULT_CHANNEL_ID = "com.dieam.reactnativepushnotification.default_notification_channel_id";
+    private static final String KEY_NOTIFICATION_FOREGROUND = "com.dieam.reactnativepushnotification.notification_foreground";
     private static final String KEY_NOTIFICATION_COLOR = "com.dieam.reactnativepushnotification.notification_color";
 
     private static Bundle metadata;
@@ -30,24 +30,21 @@ class RNPushNotificationConfig {
         }
     }
 
-    public String getChannelName() {
+    private String getStringValue(String key, String defaultValue) {
         try {
-            return metadata.getString(KEY_CHANNEL_NAME);
+            final String value = metadata.getString(key);
+
+            if (value != null && value.length() > 0) {
+                return value;
+            }
         } catch (Exception e) {
-            Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_CHANNEL_NAME + " in manifest. Falling back to default");
+            Log.w(RNPushNotification.LOG_TAG, "Unable to find " + key + " in manifest. Falling back to default");
         }
+
         // Default
-        return "rn-push-notification-channel";
+        return defaultValue;
     }
-    public String getChannelDescription() {
-        try {
-            return metadata.getString(KEY_CHANNEL_DESCRIPTION);
-        } catch (Exception e) {
-            Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_CHANNEL_DESCRIPTION + " in manifest. Falling back to default");
-        }
-        // Default
-        return "";
-    }
+
     public int getNotificationColor() {
         try {
             int resourceId = metadata.getInt(KEY_NOTIFICATION_COLOR);
@@ -59,35 +56,25 @@ class RNPushNotificationConfig {
         return -1;
     }
 
-    public String getChannelPrefix() {
+    public boolean getNotificationForeground() {
         try {
-            return metadata.getString(KEY_CHANNEL_PREFIX);
+            return metadata.getBoolean(KEY_NOTIFICATION_FOREGROUND, false);
         } catch (Exception e) {
-            Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_CHANNEL_PREFIX + " in manifest. Falling back to default");
+            Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_NOTIFICATION_FOREGROUND + " in manifest. Falling back to default");
         }
         // Default
-        return "rn-push-notification-channel";
+        return false;
     }
-    public String getChannelNameForId(String channelId) {
-        if (channelId != null) {
-            try {
-                return metadata.getString(KEY_CHANNEL_NAME + "." + channelId);
-            } catch (Exception e) {
-                Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_CHANNEL_NAME + "." + channelId + " in manifest. Falling back to default");
-            }
+
+    public String getNotificationDefaultChannelId() {
+        try {
+            return getStringValue(KEY_NOTIFICATION_DEFAULT_CHANNEL_ID,
+              getStringValue(KEY_NOTIFICATION_FIREBASE_DEFAULT_CHANNEL_ID, "fcm_fallback_notification_channel")
+            );
+        } catch (Exception e) {
+            Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_NOTIFICATION_DEFAULT_CHANNEL_ID + " in manifest. Falling back to default");
         }
         // Default
-        return "rn-push-notification-channel";
-    }
-    public String getChannelDescriptionForId(String channelId) {
-        if (channelId != null) {
-            try {
-                return metadata.getString(KEY_CHANNEL_DESCRIPTION + "." + channelId);
-            } catch (Exception e) {
-                Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_CHANNEL_DESCRIPTION + "." + channelId + " in manifest. Falling back to default");
-            }
-        }
-        // Default
-        return "";
+        return "fcm_fallback_notification_channel";
     }
 }
