@@ -6,41 +6,137 @@
  * @flow
  */
 
-import React, { Component } from 'react';
-import { TextInput, StyleSheet, Text, View, TouchableOpacity, Alert } from 'react-native';
+import React, {Component} from 'react';
+import {
+  TextInput,
+  StyleSheet,
+  Text,
+  View,
+  TouchableOpacity,
+  Alert,
+} from 'react-native';
 import NotifService from './NotifService';
-import appConfig from './app.json';
 
-type Props = {};
-export default class App extends Component<Props> {
-
+export default class App extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-      senderId: appConfig.senderID
-    };
+    this.state = {};
 
-    this.notif = new NotifService(this.onRegister.bind(this), this.onNotif.bind(this));
+    this.notif = new NotifService(
+      this.onRegister.bind(this),
+      this.onNotif.bind(this),
+    );
   }
 
   render() {
     return (
       <View style={styles.container}>
-        <Text style={styles.title}>Example app react-native-push-notification</Text>
+        <Text style={styles.title}>
+          Example app react-native-push-notification
+        </Text>
         <View style={styles.spacer}></View>
-        <TextInput style={styles.textField} value={this.state.registerToken} placeholder="Register token" />
+        <TextInput
+          style={styles.textField}
+          value={this.state.registerToken}
+          placeholder="Register token"
+        />
         <View style={styles.spacer}></View>
 
-        <TouchableOpacity style={styles.button} onPress={() => { this.notif.localNotif() }}><Text>Local Notification (now)</Text></TouchableOpacity>
-        <TouchableOpacity style={styles.button} onPress={() => { this.notif.scheduleNotif() }}><Text>Schedule Notification in 30s</Text></TouchableOpacity>
-        <TouchableOpacity style={styles.button} onPress={() => { this.notif.cancelNotif() }}><Text>Cancel last notification (if any)</Text></TouchableOpacity>
-        <TouchableOpacity style={styles.button} onPress={() => { this.notif.cancelAll() }}><Text>Cancel all notifications</Text></TouchableOpacity>
-        <TouchableOpacity style={styles.button} onPress={() => { this.notif.checkPermission(this.handlePerm.bind(this)) }}><Text>Check Permission</Text></TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.localNotif();
+          }}>
+          <Text>Local Notification (now)</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.localNotif('sample.mp3');
+          }}>
+          <Text>Local Notification with sound (now)</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.scheduleNotif();
+          }}>
+          <Text>Schedule Notification in 30s</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.scheduleNotif('sample.mp3');
+          }}>
+          <Text>Schedule Notification with sound in 30s</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.cancelNotif();
+          }}>
+          <Text>Cancel last notification (if any)</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.cancelAll();
+          }}>
+          <Text>Cancel all notifications</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.checkPermission(this.handlePerm.bind(this));
+          }}>
+          <Text>Check Permission</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.requestPermissions();
+          }}>
+          <Text>Request Permissions</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.abandonPermissions();
+          }}>
+          <Text>Abandon Permissions</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.getScheduledLocalNotifications(notifs => console.log(notifs));
+          }}>
+          <Text>Console.Log Scheduled Local Notifications</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.getDeliveredNotifications(notifs => console.log(notifs));
+          }}>
+          <Text>Console.Log Delivered Notifications</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.createOrUpdateChannel();
+          }}>
+          <Text>Create or update a channel</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => {
+            this.notif.popInitialNotification();
+          }}>
+          <Text>popInitialNotification</Text>
+        </TouchableOpacity>
 
         <View style={styles.spacer}></View>
-        <TextInput style={styles.textField} value={this.state.senderId} onChangeText={(e) => {this.setState({ senderId: e })}} placeholder="GCM ID" />
-        <TouchableOpacity style={styles.button} onPress={() => { this.notif.configure(this.onRegister.bind(this), this.onNotif.bind(this), this.state.senderId) }}><Text>Configure Sender ID</Text></TouchableOpacity>
-        {this.state.gcmRegistered && <Text>GCM Configured !</Text>}
+
+        {this.state.fcmRegistered && <Text>FCM Configured !</Text>}
 
         <View style={styles.spacer}></View>
       </View>
@@ -48,21 +144,17 @@ export default class App extends Component<Props> {
   }
 
   onRegister(token) {
-    Alert.alert("Registered !", JSON.stringify(token));
-    console.log(token);
-    this.setState({ registerToken: token.token, gcmRegistered: true });
+    this.setState({registerToken: token.token, fcmRegistered: true});
   }
 
   onNotif(notif) {
-    console.log(notif);
     Alert.alert(notif.title, notif.message);
   }
 
   handlePerm(perms) {
-    Alert.alert("Permissions", JSON.stringify(perms));
+    Alert.alert('Permissions', JSON.stringify(perms));
   }
 }
-
 
 const styles = StyleSheet.create({
   container: {
@@ -78,26 +170,26 @@ const styles = StyleSheet.create({
   },
   button: {
     borderWidth: 1,
-    borderColor: "#000000",
+    borderColor: '#000000',
     margin: 5,
     padding: 5,
-    width: "70%",
-    backgroundColor: "#DDDDDD",
+    width: '70%',
+    backgroundColor: '#DDDDDD',
     borderRadius: 5,
   },
   textField: {
     borderWidth: 1,
-    borderColor: "#AAAAAA",
+    borderColor: '#AAAAAA',
     margin: 5,
     padding: 5,
-    width: "70%"
+    width: '70%',
   },
   spacer: {
     height: 10,
   },
   title: {
-    fontWeight: "bold",
+    fontWeight: 'bold',
     fontSize: 20,
-    textAlign: "center",
-  }
+    textAlign: 'center',
+  },
 });
